+++ src/Resources/BaseDbTableResource.php
@@ -493,7 +493,7 @@ abstract class BaseDbTableResource extends BaseDbResource
 
         $records = ResourcesWrapper::unwrapResources($this->getPayloadData());
         if (empty($records)) {
-            throw new BadRequestException('No record(s) detected in request.');
+            throw new BadRequestException('No record(s) detected in request.' . ResourcesWrapper::getWrapperMsg());
         }
 
         $options = $this->request->getParameters();
@@ -539,7 +539,7 @@ abstract class BaseDbTableResource extends BaseDbResource
 
         $records = ResourcesWrapper::unwrapResources($this->getPayloadData());
         if (empty($records)) {
-            throw new BadRequestException('No record(s) detected in request.');
+            throw new BadRequestException('No record(s) detected in request.' . ResourcesWrapper::getWrapperMsg());
         }
 
         $ids = array_get($options, ApiOptions::IDS);
@@ -605,7 +605,7 @@ abstract class BaseDbTableResource extends BaseDbResource
 
         $records = ResourcesWrapper::unwrapResources($this->getPayloadData());
         if (empty($records)) {
-            throw new BadRequestException('No record(s) detected in request.');
+            throw new BadRequestException('No record(s) detected in request.' . ResourcesWrapper::getWrapperMsg());
         }
 
         $ids = array_get($options, ApiOptions::IDS);